move validation to validation package
